fix table dissapearing

+++ warehouse/src/main/java/com/example/warehouse/repositories/WarehouseRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.example.warehouse.models.Warehouse;
 
-public interface WarehouseRepository extends JpaRepository<Warehouse, Long> {
+public interface WarehouseRepository extends JpaRepository<Warehouse, Integer> {
 }
